refactor(server): migrate model building and creation to standalone schema
