+++ lib/index.js
@@ -1,7 +1,14 @@
-import Factory from "./factory";
-import { IdentityManager, Model, Collection, HasMany, BelongsTo } from "./orm";
-import association from "./association";
-import trait from "./trait";
+import {
+  Collection,
+  Factory,
+  IdentityManager,
+  Model,
+  Schema,
+  association,
+  belongsTo,
+  hasMany,
+  trait,
+} from "./orm";
 import Response from "./response";
 import Server, { createServer } from "./server";
 import Serializer from "./serializer";
@@ -45,26 +52,13 @@ import _utilsUuid from "./utils/uuid";
 
 import PretenderInterceptor from "./mock-server/pretender-config";
 
-/**
-  @hide
-*/
-function hasMany(...args) {
-  return new HasMany(...args);
-}
-
-/**
-  @hide
-*/
-function belongsTo(...args) {
-  return new BelongsTo(...args);
-}
-
 export {
   association,
   trait,
   Model,
   Collection,
   Serializer,
+  Schema,
   ActiveModelSerializer,
   JSONAPISerializer,
   RestSerializer,
