+++ lib/orm/associations/belongs-to.js
@@ -302,3 +302,7 @@ export default class BelongsTo extends Association {
     });
   }
 }
+
+export function belongsTo(...args) {
+  return new BelongsTo(...args);
+}
