+++ lib/orm/associations/has-many.js
@@ -345,3 +345,7 @@ export default class HasMany extends Association {
     });
   }
 }
+
+export function hasMany(...args) {
+  return new HasMany(...args);
+}
