+++ lib/orm/db/db.js
@@ -1,6 +1,9 @@
 import DbCollection from "./db-collection";
 import IdentityManager from "./identity-manager";
 import cloneDeep from "lodash/cloneDeep.js";
+import { singularize, pluralize } from "inflected";
+
+const defaultInflector = { singularize, pluralize };
 
 /**
   Your Mirage server has a database which you can interact with in your route handlers. You’ll typically use models to interact with your database data, but you can always reach into the db directly in the event you want more control.
@@ -18,10 +21,15 @@ import cloneDeep from "lodash/cloneDeep.js";
   @public
  */
 class Db {
-  constructor(initialData, identityManagers) {
+  constructor(
+    initialData,
+    identityManagers = {},
+    inflector = defaultInflector
+  ) {
     this._collections = [];
 
     this.registerIdentityManagers(identityManagers);
+    this.registerInflector(inflector);
 
     if (initialData) {
       this.loadData(initialData);
@@ -170,7 +178,7 @@ class Db {
    */
   identityManagerFor(name) {
     return (
-      this._identityManagers[this._container.inflector.singularize(name)] ||
+      this._identityManagers[this._inflector.singularize(name)] ||
       this._identityManagers.application ||
       IdentityManager
     );
@@ -182,7 +190,17 @@ class Db {
     @hide
    */
   registerIdentityManagers(identityManagers) {
-    this._identityManagers = identityManagers || {};
+    this._identityManagers = identityManagers;
+  }
+
+  /**
+   * @method registerInflector
+   * @param {Object} inflector
+   * @public
+   * @hide
+   */
+  registerInflector(inflector) {
+    this._inflector = inflector;
   }
 }
