+++ lib/orm/factory/factory.js
@@ -1,7 +1,7 @@
-import isPlainObject from "lodash/isPlainObject.js";
-import isFunction from "lodash/isFunction.js";
-import mapValues from "lodash/mapValues.js";
-import referenceSort from "./utils/reference-sort";
+import isPlainObject from "lodash/isPlainObject";
+import isFunction from "lodash/isFunction";
+import mapValues from "lodash/mapValues";
+import referenceSort from "../../utils/reference-sort";
 
 let Factory = function () {
   this.build = function (sequence) {
