+++ lib/orm/index.js
@@ -1,8 +1,14 @@
 // Core ORM components
-export { default as Schema } from "./schema";
-export { default as Model } from "./model";
 export { default as Collection } from "./collection";
+export { default as Model } from "./model";
 export { default as PolymorphicCollection } from "./polymorphic-collection";
+export { default as Schema } from "./schema";
+
+// Factory components
+export { default as Factory } from "./factory/factory";
+export { default as FactoryManager } from "./factory/factoryManager";
+export { default as association } from "./factory/association";
+export { default as trait } from "./factory/trait";
 
 // DB components
 export { default as Db } from "./db/db";
@@ -11,5 +17,5 @@ export { default as IdentityManager } from "./db/identity-manager";
 
 // Association types
 export { default as Association } from "./associations/association";
-export { default as BelongsTo } from "./associations/belongs-to";
-export { default as HasMany } from "./associations/has-many";
+export { default as BelongsTo, belongsTo } from "./associations/belongs-to";
+export { default as HasMany, hasMany } from "./associations/has-many";
