+++ lib/orm/schema.js
@@ -1,12 +1,21 @@
+import forIn from "lodash/forIn";
+import find from "lodash/find";
+import isInteger from "lodash/isInteger";
+import isPlainObject from "lodash/isPlainObject";
+import { singularize, pluralize } from "inflected";
+import assert from "../assert";
 import { camelize, dasherize } from "../utils/inflector";
 import Association from "./associations/association";
 import Collection from "./collection";
-import assert from "../assert";
-import forIn from "lodash/forIn.js";
+import Db from "./db/db";
+import FactoryManager from "./factory/factoryManager";
 
 const collectionNameCache = {};
 const internalCollectionNameCache = {};
 const modelNameCache = {};
+const isPluralForModelCache = {};
+
+const defaultInflector = { singularize, pluralize };
 
 /**
   The primary use of the `Schema` class is to use it to find Models and Collections via the `Model` class methods.
@@ -32,20 +41,50 @@ const modelNameCache = {};
   @public
  */
 export default class Schema {
-  constructor(db, modelsMap = {}) {
-    assert(db, "A schema requires a db");
+  constructor(config = {}) {
+    this.inflector = config.inflector || defaultInflector;
 
     /**
-      Returns Mirage's database. See the `Db` docs for the db's API.
-
+      Creates a new Mirage database. See the `Db` docs for the db's API.
       @property db
       @type {Object}
       @public
     */
-    this.db = db;
+    this.db = new Db(
+      config.initialData,
+      config.identityManagers,
+      this.inflector
+    );
+
+    /**
+     * Registry of models and their associations
+     * @type {Object}
+     * @private
+     */
     this._registry = {};
+
+    /**
+     * Registry of polymorphic associations
+     * @type {Object}
+     * @private
+     */
     this._dependentAssociations = { polymorphic: [] };
-    this.registerModels(modelsMap);
+
+    /**
+     * Factory builder
+     * @type {FactoryManager}
+     * @private
+     */
+    this._factoryManager = new FactoryManager(this);
+
+    if (config.models) {
+      this.registerModels(config.models);
+    }
+
+    if (config.factories) {
+      this._factoryManager.registerFactories(config.factories);
+    }
+
     this.isSaving = {}; // a hash of models that are being saved, used to avoid cycles
   }
 
@@ -56,8 +95,8 @@ export default class Schema {
     @hide
    */
   registerModels(hash = {}) {
-    forIn(hash, (model, key) => {
-      this.registerModel(key, hash[key]);
+    forIn(hash, (modelClass, key) => {
+      this.registerModel(key, modelClass);
     });
   }
 
@@ -157,6 +196,155 @@ export default class Schema {
   }
 
   /**
+   * @method factoryFor
+   * @param type
+   * @public
+   */
+  factoryFor(type) {
+    return this._factoryManager.factoryFor(type);
+  }
+
+  /**
+    @method registerFactories
+    @param factoryMap
+    @public
+    @hide
+   */
+  registerFactories(factoryMap) {
+    this._factoryManager.registerFactories(factoryMap);
+  }
+
+  /**
+    @method build
+    @param type
+    @param traitsAndOverrides
+    @public
+   */
+  build(type, ...options) {
+    return this._factoryManager.build(type, ...options);
+  }
+
+  /**
+   * @method buildList
+   * @param type
+   * @param amount
+   * @param traitsAndOverrides
+   * @public
+   */
+  buildList(type, amount, ...options) {
+    return this._factoryManager.buildList(type, amount, ...options);
+  }
+
+  /**
+    Creates a new model instance with attributes *attrs*, including any overrides from *traits*, and inserts it into the database.
+
+    ```js
+    let post = schema.create('post', { title: 'Lorem ipsum' }, 'published');
+    post.title;     // Lorem ipsum
+    post.id;        // 1
+    post.published; // true
+    post.isNew();   // false
+    ```
+
+    @method create
+    @param type the singularized type of the model
+    @param traitsAndOverrides an object of attributes and traits to override
+    @public
+  */
+  create(type, ...options) {
+    assert(
+      this._modelOrFactoryExistsForType(type),
+      `You called create('${type}') but no model or factory was found. Make sure you're passing in the singularized version of the model or factory name.`
+    );
+
+    // When there is a Model defined, we should return an instance
+    // of it instead of returning the bare attributes.
+    let traits = options.filter((arg) => arg && typeof arg === "string");
+    let overrides = find(options, (arg) => isPlainObject(arg));
+    let collectionFromCreateList = find(
+      options,
+      (arg) => arg && Array.isArray(arg)
+    );
+
+    let attrs = this.build(type, ...traits, overrides);
+    let modelOrRecord;
+
+    if (this[this.toCollectionName(type)]) {
+      let modelClass = this[this.toCollectionName(type)];
+
+      modelOrRecord = modelClass.new(attrs).save();
+    } else {
+      let collection, collectionName;
+
+      if (collectionFromCreateList) {
+        collection = collectionFromCreateList;
+      } else {
+        collectionName = this.toInternalCollectionName(type);
+        collection = this.db[collectionName];
+      }
+
+      assert(
+        collection,
+        `You called create('${type}') but no model or factory was found.`
+      );
+      modelOrRecord = collection.insert(attrs);
+    }
+
+    let OriginalFactory = this.factoryFor(type);
+    if (OriginalFactory) {
+      OriginalFactory.extractAfterCreateCallbacks({ traits }).forEach(
+        (afterCreate) => {
+          afterCreate(modelOrRecord, this);
+        }
+      );
+    }
+
+    return modelOrRecord;
+  }
+
+  /**
+    Creates *amount* models of type *type*, optionally overriding the attributes from the factory with *attrs*.
+
+    @method createList
+    @param type the singularized type of the model
+    @param amount the number of models to create
+    @param traitsAndOverrides an object of attributes and traits to override
+    @public
+
+    ```js
+    let posts = schema.createList('post', 5, { title: 'Lorem ipsum' }, 'published');
+    posts.length;       // 5
+    posts[0].title;     // Lorem ipsum
+    posts[0].id;        // 1
+    posts[0].published; // true
+    ```
+  */
+  createList(type, amount, ...traitsAndOverrides) {
+    assert(
+      this._modelOrFactoryExistsForType(type),
+      `You called createList('${type}') but no model or factory was found. Make sure you're passing in the singularized version of the model or factory name.`
+    );
+    assert(
+      isInteger(amount),
+      `second argument has to be an integer, you passed: ${typeof amount}`
+    );
+
+    let list = [];
+    let collectionName =
+      this.toInternalCollectionName(type) ||
+      `_${this.inflector.pluralize(type)}`;
+    let collection = this.db[collectionName];
+
+    const createArguments = [type, ...traitsAndOverrides, collection];
+
+    for (let i = 0; i < amount; i++) {
+      list.push(this.create.apply(this, createArguments));
+    }
+
+    return list;
+  }
+
+  /**
     Create a new unsaved model instance with attributes *attrs*.
 
     ```js
@@ -173,25 +361,6 @@ export default class Schema {
    */
   new(type, attrs) {
     return this._instantiateModel(dasherize(type), attrs);
-  }
-
-  /**
-    Create a new model instance with attributes *attrs*, and insert it into the database.
-
-    ```js
-    let post = blogPosts.create({title: 'Lorem ipsum'});
-    post.title;   // Lorem ipsum
-    post.id;      // 1
-    post.isNew(); // false
-    ```
-
-    @method create
-    @param type
-    @param attrs
-    @public
-   */
-  create(type, attrs) {
-    return this.new(type, attrs).save();
   }
 
   /**
@@ -467,7 +636,7 @@ export default class Schema {
   }
 
   hasModelForModelName(modelName) {
-    return this.modelFor(camelize(modelName));
+    return !!this.modelFor(camelize(modelName));
   }
 
   /*
@@ -494,9 +663,7 @@ export default class Schema {
     if (typeof collectionNameCache[type] !== "string") {
       let modelName = dasherize(type);
 
-      const collectionName = camelize(
-        this._container.inflector.pluralize(modelName)
-      );
+      const collectionName = camelize(this.inflector.pluralize(modelName));
 
       collectionNameCache[type] = collectionName;
     }
@@ -520,7 +687,7 @@ export default class Schema {
     if (typeof modelNameCache[type] !== "string") {
       let dasherized = dasherize(type);
 
-      const modelName = this._container.inflector.singularize(dasherized);
+      const modelName = this.inflector.singularize(dasherized);
 
       modelNameCache[type] = modelName;
     }
@@ -602,5 +769,51 @@ export default class Schema {
     } else {
       return null;
     }
+  }
+
+  _typeIsPluralForModel(typeOrCollectionName) {
+    if (typeof isPluralForModelCache[typeOrCollectionName] !== "boolean") {
+      let modelOrFactoryExists =
+        this._modelOrFactoryExistsForTypeOrCollectionName(typeOrCollectionName);
+      let isPlural =
+        typeOrCollectionName === this.inflector.pluralize(typeOrCollectionName);
+      let isUncountable =
+        this.inflector.singularize(typeOrCollectionName) ===
+        this.inflector.pluralize(typeOrCollectionName);
+
+      const isPluralForModel =
+        isPlural && !isUncountable && modelOrFactoryExists;
+
+      isPluralForModelCache[typeOrCollectionName] = isPluralForModel;
+    }
+
+    return isPluralForModelCache[typeOrCollectionName];
+  }
+
+  /**
+   *
+   * @private
+   * @hide
+   */
+  _modelOrFactoryExistsForType(type) {
+    let modelExists = this.hasModelForModelName(type);
+    let dbCollectionExists = this.db[this.toInternalCollectionName(type)];
+
+    return (
+      (modelExists || dbCollectionExists) && !this._typeIsPluralForModel(type)
+    );
+  }
+
+  /**
+   *
+   * @private
+   * @hide
+   */
+  _modelOrFactoryExistsForTypeOrCollectionName(typeOrCollectionName) {
+    let modelExists = this.hasModelForModelName(typeOrCollectionName);
+    let dbCollectionExists =
+      this.db[this.toInternalCollectionName(typeOrCollectionName)];
+
+    return modelExists || dbCollectionExists;
   }
 }
