+++ lib/server.js
@@ -1,22 +1,11 @@
 /* eslint no-console: 0 */
 
+import get from "lodash/get";
+import pick from "lodash/pick";
 import { camelize } from "./utils/inflector";
-import isAssociation from "./utils/is-association";
 import assert from "./assert";
-import { BelongsTo } from "./orm";
 import Container from "./container";
-import { singularize, pluralize } from "inflected";
-import pick from "lodash/pick.js";
-import assign from "lodash/assign.js";
-import find from "lodash/find.js";
-import isPlainObject from "lodash/isPlainObject.js";
-import isInteger from "lodash/isInteger.js";
-
 import PretenderConfig from "./mock-server/pretender-config";
-
-const isPluralForModelCache = {};
-
-const defaultInflector = { singularize, pluralize };
 
 /**
  * Creates a Server
@@ -76,23 +65,72 @@ export default class Server {
    */
   constructor(options = {}) {
     this._container = new Container();
+
+    /**
+     * Sets up the server with the given options.
+     * @param {Object} options - The options to configure the server with.
+     */
     this.config(options);
 
     /**
-      Returns the Mirage Db instance.
+      Returns the Mirage Schema (ORM) instance.
+      @property schema
+      @return Schema
+      @public
+    */
+    this.schema = this.schema || undefined;
 
+    /**
+      Returns the Mirage Database instance.
       @property db
-      @return Db
+      @return Database
+      @public
     */
     this.db = this.db || undefined;
 
     /**
-      Returns the Mirage Schema (ORM) instance.
+      Mirage needs know the singular and plural versions of certain words for some of its APIs to work.
 
-      @property schema
-      @return Schema
+      For example, whenever you define a model
+
+      ```js
+      createServer({
+        models: {
+          post: Model
+        }
+      })
+      ```
+
+      Mirage will pluralize the word "post" and use it to create a `db.posts` database collection.
+
+      To accomplish this, Mirage uses an object called an Inflector. An Inflector is an object with two methods, `singularize` and `pluralize`, that Mirage will call whenever it needs to inflect a word.
+
+      Mirage has a default inflector, so if you write
+
+      ```js
+      createServer()
+      ```
+
+      you'll be using the node [inflected](https://github.com/martinandert/inflected#readme) package. This can be customized if you have irregular words or need to change the defaults. You can wead more in [the guide on customizing inflections](/docs/advanced/customizing-inflections).
+
+      You typically should be able to make your customizations using the provided inflector. It's good to match any custom inflections your backend uses, as this will keep your Mirage code more consistent and simpler.
+
+      You can also override the inflector completely and provide your own `pluralize` and `singularize` methods:
+
+      ```js
+      createServer({
+        inflector: {
+          pluralize(word) {
+            // your logic
+          },
+          singularize(word) {
+            // your logic
+          }
+        }
+      })
+      ```
     */
-    this.schema = this.schema || undefined;
+    this.inflector = this.inflector || undefined;
 
     this.middleware = [];
   }
@@ -170,51 +208,6 @@ export default class Server {
     this._config = config;
 
     /**
-      Mirage needs know the singular and plural versions of certain words for some of its APIs to work.
-
-      For example, whenever you define a model
-
-      ```js
-      createServer({
-        models: {
-          post: Model
-        }
-      })
-      ```
-
-      Mirage will pluralize the word "post" and use it to create a `db.posts` database collection.
-
-      To accomplish this, Mirage uses an object called an Inflector. An Inflector is an object with two methods, `singularize` and `pluralize`, that Mirage will call whenever it needs to inflect a word.
-
-      Mirage has a default inflector, so if you write
-
-      ```js
-      createServer()
-      ```
-
-      you'll be using the node [inflected](https://github.com/martinandert/inflected#readme) package. This can be customized if you have irregular words or need to change the defaults. You can wead more in [the guide on customizing inflections](/docs/advanced/customizing-inflections).
-
-      You typically should be able to make your customizations using the provided inflector. It's good to match any custom inflections your backend uses, as this will keep your Mirage code more consistent and simpler.
-
-      You can also override the inflector completely and provide your own `pluralize` and `singularize` methods:
-
-      ```js
-      createServer({
-        inflector: {
-          pluralize(word) {
-            // your logic
-          },
-          singularize(word) {
-            // your logic
-          }
-        }
-      })
-      ```
-    */
-    this.inflector = config.inflector || defaultInflector;
-    this._container.register("inflector", this.inflector);
-
-    /**
       Set to `true` or `false` to explicitly specify logging behavior.
 
       By default, server responses are logged in non-testing environments. Logging is disabled by default in testing, so as not to clutter CI test runner output.
@@ -256,21 +249,14 @@ export default class Server {
 
     this.trackRequests = config.trackRequests;
 
-    if (this.db) {
-      this.db.registerIdentityManagers(config.identityManagers);
-    } else {
-      this.db = this._container.create(
-        "Db",
-        undefined,
-        config.identityManagers
-      );
-    }
-
     if (this.schema) {
       this.schema.registerModels(config.models);
+      this.db.registerIdentityManagers(config.identityManagers);
       this.serializerOrRegistry.registerSerializers(config.serializers || {});
     } else {
-      this.schema = this._container.create("Schema", this.db, config.models);
+      this.schema = this._container.create("Schema", config);
+      this.db = this.schema.db;
+      this.inflector = this.schema.inflector;
       this.serializerOrRegistry = this._container.create(
         "SerializerRegistry",
         this.schema,
@@ -278,10 +264,12 @@ export default class Server {
       );
     }
 
-    let hasFactories = this._hasModulesOfType(config, "factories");
+    let factories = get(config, "factories");
+    let hasFactories = !!factories;
+
+    let defaultScenario = get(config, "scenarios.default");
     let hasDefaultScenario =
-      config.scenarios &&
-      Object.prototype.hasOwnProperty.call(config.scenarios, "default");
+      defaultScenario && typeof defaultScenario === "function";
 
     if (config.baseConfig) {
       this.loadConfig(config.baseConfig);
@@ -296,10 +284,10 @@ export default class Server {
     }
 
     if (this.isTest() && hasFactories) {
-      this.loadFactories(config.factories);
+      this.schema.registerFactories(factories);
     } else if (!this.isTest() && hasDefaultScenario) {
-      this.loadFactories(config.factories);
-      config.scenarios.default(this);
+      this.schema.registerFactories(factories);
+      defaultScenario(this);
     } else {
       this.loadFixtures();
     }
@@ -417,91 +405,33 @@ export default class Server {
     this.db.loadData(fixtures);
   }
 
-  /*
-    Factory methods
-  */
-
   /**
-   * Load factories into Mirage's database.
+   * Loads factories into the server.
+   *
+   * Usually you will not need to call this method directly, as factories will be loaded automatically when the server is created, like this:
+   * ```js
+   * createServer({
+   *   factories: {
+   *     contact: Factory.extend({
+   *       name: "Link",
+   *     }),
+   *   },
+   * });
+   * ```
+   *
+   * However, if you need to load factories from a file, or if you need to load factories after the server has been created, you can do so like this:
+   *
+   * ```js
+   * import factories from './factories';
+   * server.loadFactories(factories);
+   * ```
    *
    * @method loadFactories
-   * @param {Object} factoryMap
+   * @param {Object} factoryMap - A map of factory names to factory definitions.
    * @public
-   * @hide
    */
-  loadFactories(factoryMap = {}) {
-    // Store a reference to the factories
-    let currentFactoryMap = this._factoryMap || {};
-    this._factoryMap = assign(currentFactoryMap, factoryMap);
-
-    // Create a collection for each factory
-    Object.keys(factoryMap).forEach((type) => {
-      let collectionName = this.schema.toCollectionName(type);
-      this.db.createCollection(collectionName);
-    });
-  }
-
-  /**
-   * Get the factory for a given type.
-   *
-   * @method factoryFor
-   * @param {String} type
-   * @private
-   * @hide
-   */
-  factoryFor(type) {
-    let camelizedType = camelize(type);
-
-    if (this._factoryMap && this._factoryMap[camelizedType]) {
-      return this._factoryMap[camelizedType];
-    }
-  }
-
-  build(type, ...traitsAndOverrides) {
-    let traits = traitsAndOverrides.filter(
-      (arg) => arg && typeof arg === "string"
-    );
-    let overrides = find(traitsAndOverrides, (arg) => isPlainObject(arg));
-    let camelizedType = camelize(type);
-
-    // Store sequence for factory type as instance variable
-    this.factorySequences = this.factorySequences || {};
-    this.factorySequences[camelizedType] =
-      this.factorySequences[camelizedType] + 1 || 0;
-
-    let OriginalFactory = this.factoryFor(type);
-    if (OriginalFactory) {
-      OriginalFactory = OriginalFactory.extend({});
-      let attrs = OriginalFactory.attrs || {};
-      this._validateTraits(traits, OriginalFactory, type);
-      let mergedExtensions = this._mergeExtensions(attrs, traits, overrides);
-      this._mapAssociationsFromAttributes(type, attrs, overrides);
-      this._mapAssociationsFromAttributes(type, mergedExtensions);
-
-      let Factory = OriginalFactory.extend(mergedExtensions);
-      let factory = new Factory();
-
-      let sequence = this.factorySequences[camelizedType];
-      return factory.build(sequence);
-    } else {
-      return overrides;
-    }
-  }
-
-  buildList(type, amount, ...traitsAndOverrides) {
-    assert(
-      isInteger(amount),
-      `second argument has to be an integer, you passed: ${typeof amount}`
-    );
-
-    let list = [];
-
-    const buildArgs = [type, ...traitsAndOverrides];
-    for (let i = 0; i < amount; i++) {
-      list.push(this.build.apply(this, buildArgs));
-    }
-
-    return list;
+  loadFactories(factoryMap) {
+    this.schema.registerFactories(factoryMap);
   }
 
   /**
@@ -545,60 +475,11 @@ export default class Server {
 
     @method create
     @param type the singularized type of the model
-    @param traitsAndOverrides
+    @param traitsAndOverrides an object of attributes and traits to override
     @public
   */
-  create(type, ...options) {
-    assert(
-      this._modelOrFactoryExistsForType(type),
-      `You called server.create('${type}') but no model or factory was found. Make sure you're passing in the singularized version of the model or factory name.`
-    );
-
-    // When there is a Model defined, we should return an instance
-    // of it instead of returning the bare attributes.
-    let traits = options.filter((arg) => arg && typeof arg === "string");
-    let overrides = find(options, (arg) => isPlainObject(arg));
-    let collectionFromCreateList = find(
-      options,
-      (arg) => arg && Array.isArray(arg)
-    );
-
-    let attrs = this.build(type, ...traits, overrides);
-    let modelOrRecord;
-
-    if (this.schema && this.schema[this.schema.toCollectionName(type)]) {
-      let modelClass = this.schema[this.schema.toCollectionName(type)];
-
-      modelOrRecord = modelClass.create(attrs);
-    } else {
-      let collection, collectionName;
-
-      if (collectionFromCreateList) {
-        collection = collectionFromCreateList;
-      } else {
-        collectionName = this.schema
-          ? this.schema.toInternalCollectionName(type)
-          : `_${this.inflector.pluralize(type)}`;
-        collection = this.db[collectionName];
-      }
-
-      assert(
-        collection,
-        `You called server.create('${type}') but no model or factory was found.`
-      );
-      modelOrRecord = collection.insert(attrs);
-    }
-
-    let OriginalFactory = this.factoryFor(type);
-    if (OriginalFactory) {
-      OriginalFactory.extractAfterCreateCallbacks({ traits }).forEach(
-        (afterCreate) => {
-          afterCreate(modelOrRecord, this);
-        }
-      );
-    }
-
-    return modelOrRecord;
+  create(...args) {
+    return this.schema.create(...args);
   }
 
   /**
@@ -640,29 +521,8 @@ export default class Server {
     @param traitsAndOverrides
     @public
   */
-  createList(type, amount, ...traitsAndOverrides) {
-    assert(
-      this._modelOrFactoryExistsForType(type),
-      `You called server.createList('${type}') but no model or factory was found. Make sure you're passing in the singularized version of the model or factory name.`
-    );
-    assert(
-      isInteger(amount),
-      `second argument has to be an integer, you passed: ${typeof amount}`
-    );
-
-    let list = [];
-    let collectionName = this.schema
-      ? this.schema.toInternalCollectionName(type)
-      : `_${this.inflector.pluralize(type)}`;
-    let collection = this.db[collectionName];
-
-    const createArguments = [type, ...traitsAndOverrides, collection];
-
-    for (let i = 0; i < amount; i++) {
-      list.push(this.create.apply(this, createArguments));
-    }
-
-    return list;
+  createList(...args) {
+    return this.schema.createList(...args);
   }
 
   /**
@@ -761,147 +621,5 @@ export default class Server {
         return [code, headers, this._serialize(response)];
       });
     };
-  }
-
-  /**
-   *
-   * @private
-   * @hide
-   */
-  _hasModulesOfType(modules, type) {
-    let modulesOfType = modules[type];
-    return modulesOfType ? Object.keys(modulesOfType).length > 0 : false;
-  }
-
-  /**
-   *
-   * @private
-   * @hide
-   */
-  _typeIsPluralForModel(typeOrCollectionName) {
-    if (typeof isPluralForModelCache[typeOrCollectionName] !== "boolean") {
-      let modelOrFactoryExists =
-        this._modelOrFactoryExistsForTypeOrCollectionName(typeOrCollectionName);
-      let isPlural =
-        typeOrCollectionName === this.inflector.pluralize(typeOrCollectionName);
-      let isUncountable =
-        this.inflector.singularize(typeOrCollectionName) ===
-        this.inflector.pluralize(typeOrCollectionName);
-
-      const isPluralForModel =
-        isPlural && !isUncountable && modelOrFactoryExists;
-
-      isPluralForModelCache[typeOrCollectionName] = isPluralForModel;
-    }
-
-    return isPluralForModelCache[typeOrCollectionName];
-  }
-
-  /**
-   *
-   * @private
-   * @hide
-   */
-  _modelOrFactoryExistsForType(type) {
-    let modelExists = this.schema && this.schema.modelFor(camelize(type));
-    let dbCollectionExists =
-      this.db[this.schema.toInternalCollectionName(type)];
-
-    return (
-      (modelExists || dbCollectionExists) && !this._typeIsPluralForModel(type)
-    );
-  }
-
-  /**
-   *
-   * @private
-   * @hide
-   */
-  _modelOrFactoryExistsForTypeOrCollectionName(typeOrCollectionName) {
-    let modelExists =
-      this.schema && this.schema.modelFor(camelize(typeOrCollectionName));
-    let dbCollectionExists =
-      this.db[this.schema.toInternalCollectionName(typeOrCollectionName)];
-
-    return modelExists || dbCollectionExists;
-  }
-
-  /**
-   *
-   * @private
-   * @hide
-   */
-  _validateTraits(traits, factory, type) {
-    traits.forEach((traitName) => {
-      if (!factory.isTrait(traitName)) {
-        throw new Error(
-          `'${traitName}' trait is not registered in '${type}' factory`
-        );
-      }
-    });
-  }
-
-  /**
-   *
-   * @private
-   * @hide
-   */
-  _mergeExtensions(attrs, traits, overrides) {
-    let allExtensions = traits.map((traitName) => {
-      return attrs[traitName].extension;
-    });
-    allExtensions.push(overrides || {});
-    return allExtensions.reduce((accum, extension) => {
-      return assign(accum, extension);
-    }, {});
-  }
-
-  /**
-   *
-   * @private
-   * @hide
-   */
-  _mapAssociationsFromAttributes(modelName, attributes, overrides = {}) {
-    Object.keys(attributes || {})
-      .filter((attr) => {
-        return isAssociation(attributes[attr]);
-      })
-      .forEach((attr) => {
-        let modelClass = this.schema.modelClassFor(modelName);
-        let association = modelClass.associationFor(attr);
-
-        assert(
-          association && association instanceof BelongsTo,
-          `You're using the \`association\` factory helper on the '${attr}' attribute of your ${modelName} factory, but that attribute is not a \`belongsTo\` association.`
-        );
-
-        let isSelfReferentialBelongsTo =
-          association &&
-          association instanceof BelongsTo &&
-          association.modelName === modelName;
-
-        assert(
-          !isSelfReferentialBelongsTo,
-          `You're using the association() helper on your ${modelName} factory for ${attr}, which is a belongsTo self-referential relationship. You can't do this as it will lead to infinite recursion. You can move the helper inside of a trait and use it selectively.`
-        );
-
-        let isPolymorphic =
-          association && association.opts && association.opts.polymorphic;
-
-        assert(
-          !isPolymorphic,
-          `You're using the association() helper on your ${modelName} factory for ${attr}, which is a polymorphic relationship. This is not currently supported.`
-        );
-
-        let factoryAssociation = attributes[attr];
-        let foreignKey = `${camelize(attr)}Id`;
-        if (!overrides[attr]) {
-          attributes[foreignKey] = this.create(
-            association.modelName,
-            ...factoryAssociation.traitsAndOverrides
-          ).id;
-        }
-        delete attributes[attr];
-      });
   }
 }
